order-tag 添加“添加时间”

+++ models/OrderTag.php
@@ -29,8 +29,11 @@ class OrderTag extends FaOrderTag
                 self::insertOrderTags($order->id, $order->tags);
                 break;
             case 'update' :
-                self::deleteAll(['order_id'=>$order->id]);
-                self::insertOrderTags($order->id, $order->tags);
+                # 如果新旧tag 不一致则更新
+                if (!self::checkTags($order->id, $order->tags)){
+                    self::deleteAll(['order_id'=>$order->id]);
+                    self::insertOrderTags($order->id, $order->tags);
+                }
                 break;
             case 'delete' :
                 self::deleteAll(['order_id'=>$order->id]);
@@ -56,8 +59,9 @@ class OrderTag extends FaOrderTag
         foreach ($tag_arr as $k => $v){
             $rows[$k]['order_id'] = $order_id;
             $rows[$k]['tag'] = $v;
+            $rows[$k]['add_time'] = time();
         }
-        return Yii::$app->db->createCommand()->batchInsert(OrderTag::tableName(), ['order_id', 'tag'], $rows)->execute();
+        return Yii::$app->db->createCommand()->batchInsert(OrderTag::tableName(), ['order_id', 'tag', 'add_time'], $rows)->execute();
     }
 
     /**
@@ -68,4 +72,24 @@ class OrderTag extends FaOrderTag
     {
         return self::findBySql("SELECT `tag`, COUNT(1) AS `num` FROM `xm_order_tag` GROUP BY `tag` ORDER BY `num` DESC LIMIT 10;")->asArray()->all();
     }
+
+    /**
+     * 检查原有tag和新传来的tag是否一致
+     * @param int $order_id
+     * @param string $new_tags
+     * @return bool
+     */
+    private static function checkTags(int $order_id, string $new_tags)
+    {
+        # 使用数组来比较，因为拼接字符串可能会导致因顺序不一致而差错
+        $sql = 'select tag from xm_order_tag where order_id=' . $order_id;
+        $old_tags = self::findBySql($sql)->asArray()->all();
+        $old_tags = array_column($old_tags, 'tag');
+        $new_tags = explode('|', $new_tags);
+        if ($new_tags == $old_tags){  // 使用 == 判断 即使键不一致也认为是相同的数组
+            return true;
+        }
+        return false;
+    }
+
 }
